+++ crates/mdk-core/src/prelude.rs
@@ -29,8 +29,8 @@ pub use mdk_storage_traits::GroupId;
 pub use crate::extension::NostrGroupDataExtension;
 /// Group operation results
 pub use crate::groups::{
-    GroupResult, LeafNodeDebugInfo, NostrGroupConfigData, NostrGroupDataUpdate,
-    PendingMemberChanges, RatchetTreeDebugInfo, UpdateGroupResult,
+    GroupResult, LeafNodeInfo, NostrGroupConfigData, NostrGroupDataUpdate, PendingMemberChanges,
+    RatchetTreeInfo, UpdateGroupResult,
 };
 /// Message processing result variants
 pub use crate::messages::MessageProcessingResult;